+++ mango_tango_cib/components/analysis_main.py
@@ -1,6 +1,6 @@
 from colorama import Fore
 
-from app import AnalysisContext
+from mango_tango_cib.app import AnalysisContext
 from terminal_tools import draw_box, open_directory_explorer, prompts, wait_for_key
 
 from .context import ViewContext
