+++ mango_tango_cib/components/select_project.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from app import ProjectContext
+from mango_tango_cib.app import ProjectContext
 from terminal_tools import draw_box, print_ascii_table, prompts, wait_for_key
 
 from .context import ViewContext
