+++ mango_tango_cib/components/splash.py
